requirements.txt
torch 
facenet-pytorch 
numpy
Pillow
opencv-python